requirements.txt
setuptools==50.3.2
pandas==0.25.3
redis==3.5.3
PyYAML==5.3.1
dask[complete]==2.30.0
SQLAlchemy==1.3.20
loguru==0.5.3
requests==2.24.0
numpy==1.19.2
tqdm==4.50.2
APScheduler==3.6.3
jqdatasdk==1.8.1
psycopg2==2.8.6
fastapi==0.61.1
uvicorn==0.12.2